Revert "dynamic compare page"
