+++ src/hooks/useDataPractices.ts
@@ -99,7 +99,5 @@ export default function useDataPractices() {
     return practicesByState;
   };
 
-  const practicesByState = mapPracticesByState(practicesData?.nodes ?? []);
-
-  return { practicesByState, rawPractices: practicesData };
+  return { practicesData: mapPracticesByState(practicesData?.nodes ?? []) };
 }
