+++ src/pages/compare.tsx
@@ -1,111 +1,245 @@
-import React, { useState } from 'react';
-import { Link } from 'gatsby';
+import React, { useState, useEffect } from 'react';
+import {
+  Grid,
+  GridContainer,
+  Form,
+  Fieldset,
+  Label,
+  CardGroup,
+} from '@trussworks/react-uswds';
+import useGatsbyImages from '../hooks/useGatsbyImages';
+import Select from '../components/select/Select';
+import { Option } from 'react-select';
 
 import Layout from '../components/layout/Layout';
 import Hero from '../components/hero/Hero';
 import Breadcrumbs from '../components/breadcrumbs/Breadcrumbs';
+import Card from '../components/card/Card';
 
-import useGatsbyImages from '../hooks/useGatsbyImages';
-
-import './home.scss';
 import useDataPractices from '../hooks/useDataPractices';
 
+import './home.scss';
+
 export default function Compare() {
-  const searchIcon = useGatsbyImages()['images/header/search.svg'].publicURL;
-  const [input, setInput] = useState('');
-  const OgPractices = [
-    ...new Set(
-      useDataPractices().rawPractices.nodes.map(node => node.data?.Topic),
-    ),
-  ];
-  const [practices, setPractices] = useState(OgPractices);
+  const implementedSvg =
+    useGatsbyImages()['images/compare/implementedMedium.svg'].publicURL;
+  const implementedIcon = (
+    <img className="implemented-icon" src={implementedSvg} alt="implemented" />
+  );
+  const practiceDataByState = useDataPractices().practicesData.sort((a, b) => {
+    const textA = a.name;
+    const textB = b.name;
+    return textA < textB ? -1 : textA > textB ? 1 : 0;
+  });
+  const [adminFilter, setAdminFilter] = useState<typeof Option>({
+    value: '',
+    label: 'Either',
+  });
+  const [popFilter, setPopFilter] = useState<typeof Option>({
+    value: '',
+    label: 'Any',
+  });
+  const [recFilter, setRecFilter] = useState<readonly any[]>(['']);
+  const [filteredPractices, setFilteredPractices] =
+    useState(practiceDataByState);
 
-  type alphabetRecType = {
-    letter: string;
-    practice: typeof OgPractices;
-  };
+  useEffect(() => {
+    const filterByPop = (practice: typeof practiceDataByState[0]) => {
+      const pop =
+        typeof practice.population === 'string'
+          ? parseInt(practice.population)
+          : practice.population;
+      if (popFilter.value === '0') {
+        return pop < 2500000;
+      } else if (popFilter.value === '2500000') {
+        return pop >= 2500000 && pop < 7500000;
+      } else {
+        return pop >= 7500000;
+      }
+    };
 
-  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-    e?.preventDefault();
-  };
-
-  const handleInput = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setInput(e?.currentTarget.value);
-    const filteredPractices = OgPractices.filter(
-      practice =>
-        practice
-          ?.toLowerCase()
-          .includes(e?.currentTarget.value.toLowerCase()) ?? false,
-    );
-    setPractices(filteredPractices);
-  };
-
-  const alphabet = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'.split('');
-  const reducePracticesAlphabet = () => {
-    return alphabet.reduce<alphabetRecType[]>((acc, char) => {
-      const orderedPractices = practices.filter(
-        practice => practice && practice[0].toUpperCase() === char,
+    const filterByRec = (practice: typeof practiceDataByState[0]) => {
+      const practiceArr = practice.practices
+        .filter(p => p.bool)
+        .map(p => p.practiceName);
+      return recFilter?.every(
+        filter =>
+          practiceArr.includes(filter.value) ||
+          filter.value === '' ||
+          filter === '',
       );
-      return [...acc, { letter: char, practice: orderedPractices }];
-    }, []);
+    };
+    const filteredPracticesWithUpdatedFilters = practiceDataByState
+      .sort((a, b) => {
+        const textA = a.name;
+        const textB = b.name;
+        return textA < textB ? -1 : textA > textB ? 1 : 0;
+      })
+      .filter(practice => {
+        const aFilter =
+          adminFilter === undefined ||
+          adminFilter.value === '' ||
+          adminFilter.value.includes(practice.admin);
+        const pFilter =
+          popFilter === undefined ||
+          popFilter.value === '' ||
+          filterByPop(practice);
+        const rFilter =
+          recFilter === undefined ||
+          recFilter?.length === 0 ||
+          filterByRec(practice);
+
+        return aFilter && pFilter && rFilter;
+      });
+
+    setFilteredPractices(filteredPracticesWithUpdatedFilters);
+  }, [adminFilter, popFilter, recFilter]);
+
+  const handleAdminFilter = (option: typeof Option) => {
+    setAdminFilter(option);
   };
+  const handlePopFilter = (option: typeof Option) => {
+    setPopFilter(option);
+  };
+  const handleRecFilter = (option: readonly any[]) => {
+    setRecFilter(option);
+  };
+  const adminOptions = [
+    { value: '', label: 'Either' },
+    { value: 'County', label: 'County' },
+    { value: 'State', label: 'State' },
+  ];
+  const popOptions = [
+    { value: '', label: 'Any' },
+    { value: '0', label: 'Less than 2.5 Million' },
+    { value: '2500000', label: '2.5 Million - 7.5 Million' },
+    { value: '7500000', label: 'Greater than 7.5 Million' },
+  ];
+
+  const recOptions = [
+    { value: 'No witnesses', label: 'No witnesses' },
+    { value: 'No fee', label: 'No fee' },
+    { value: 'No notary', label: 'No notary' },
+    {
+      value: 'General inbox for receiving requests',
+      label: 'General inbox for receiving requests',
+    },
+    {
+      value: 'Accepts electronic requests',
+      label: 'Accepts electronic requests',
+    },
+  ];
+
+  const createContent = (stateData: typeof practiceDataByState[0]) => (
+    <ul>
+      {stateData.practices.map(p => (
+        <li
+          key={p.practiceName}
+          className={p.bool ? 'implemented' : 'not-implemented'}
+        >
+          {p.bool ? implementedIcon : ''} {p.practiceName}
+        </li>
+      ))}
+    </ul>
+  );
 
   return (
     <Layout>
-      <section id="test-section-id">
+      <section id="hero-section">
         <Hero
           className="cwp-topic-hero"
           backgroundColor="dark"
-          title="All Practice Areas"
+          title="Compare"
         />
       </section>
-      <Breadcrumbs crumbLabel="All Practice Areas" />
-      <main className="cwp-recommendations-main">
-        <section aria-label="Big search component">
-          <form
-            className="usa-search usa-search--big"
-            role="search"
-            onSubmit={handleSubmit}
-          >
-            <label className="usa-sr-only" htmlFor="practice-search">
-              Search
-            </label>
-
-            <input
-              className="usa-input"
-              id="practice-search"
-              type="search"
-              name="search"
-              placeholder="Search practice Areas"
-              value={input}
-              onChange={handleInput}
-            />
-            <img className="search-icon" src={searchIcon} alt="search icon" />
-          </form>
-        </section>
-        <section className="recommendations-section" id="test-section-id">
-          {reducePracticesAlphabet().map(alphabetPractice => {
-            return (
-              alphabetPractice.practice.length > 0 && (
-                <section
-                  id={`${alphabetPractice.letter}-section`}
-                  key={alphabetPractice.letter}
-                >
-                  <h2>{alphabetPractice.letter}</h2>
-                  <ul>
-                    {alphabetPractice.practice.map(practice => {
-                      return (
-                        <li key={practice}>
-                          <Link to={`/compare/${practice}`}>{practice}</Link>
-                        </li>
-                      );
-                    })}
-                  </ul>
-                </section>
-              )
-            );
-          })}
-        </section>
+      <Breadcrumbs crumbLabel="Compare" />
+      <main className="cwp-main home">
+        <GridContainer>
+          <section className="intro-section">
+            <Grid>
+              <p>
+                View and compare which states and territories in the U.S. have
+                implemented our recommendations for Out of State Background
+                (Adam Walsh) Checks:{' '}
+                <strong>
+                  no witnesses, no fee, no notary, general inboxes for receiving
+                  requests, and accept electronic requests
+                </strong>
+                .
+              </p>
+              <h2 className="features-title">State by State Compare</h2>
+            </Grid>
+          </section>
+          <section className="filter-section">
+            <Form
+              onSubmit={e => {
+                e.preventDefault();
+              }}
+            >
+              <Fieldset>
+                <div className="select single">
+                  <Label htmlFor="admin">Administration</Label>
+                  <Select
+                    isMulti={false}
+                    id="admin"
+                    name="admin"
+                    selectOptions={adminOptions}
+                    value={adminFilter}
+                    handleChange={handleAdminFilter}
+                  />
+                </div>
+              </Fieldset>
+              <Fieldset>
+                <div className="select single">
+                  <Label htmlFor="pop">Population</Label>
+                  <Select
+                    isMulti={false}
+                    id="pop"
+                    name="pop"
+                    selectOptions={popOptions}
+                    value={popFilter}
+                    handleChange={handlePopFilter}
+                  />
+                </div>
+              </Fieldset>
+              <Fieldset>
+                <div className="select multi">
+                  <Label htmlFor="rec">Recommendations</Label>
+                  <Select
+                    isMulti={true}
+                    id="rec"
+                    name="rec"
+                    selectOptions={recOptions}
+                    value={recFilter}
+                    handleChange={handleRecFilter}
+                  />
+                </div>
+              </Fieldset>
+            </Form>
+          </section>
+          <section className="compare-section">
+            <Grid>
+              <p className="total">{filteredPractices.length} total results</p>
+              <CardGroup>
+                {filteredPractices.map(fp => (
+                  <Card
+                    key={fp.code}
+                    title={fp.name}
+                    content={createContent(fp)}
+                    layout="compare"
+                    className="compare-width"
+                    image={`/src/images/compare/${
+                      fp.practices.filter(p => p.bool).length
+                    }Of5.svg`}
+                    imgAlt={`${
+                      fp.practices.filter(p => p.bool).length
+                    } out of 5`}
+                  />
+                ))}
+              </CardGroup>
+            </Grid>
+          </section>
+        </GridContainer>
       </main>
     </Layout>
   );
