--- src/types/compare.ts
@@ -1,64 +1,0 @@
-type stateOptions =
-  | 'Fully Implemented'
-  | 'Partially Implemented'
-  | 'Not Implemented'
-  | 'NA'
-  | undefined;
-
-export type Practice = {
-  Name: string;
-  Topic: string;
-  AL: stateOptions;
-  AK: stateOptions;
-  AS: stateOptions;
-  AZ: stateOptions;
-  AR: stateOptions;
-  CA: stateOptions;
-  CO: stateOptions;
-  CT: stateOptions;
-  DE: stateOptions;
-  DC: stateOptions;
-  FL: stateOptions;
-  GA: stateOptions;
-  GU: stateOptions;
-  HI: stateOptions;
-  ID: stateOptions;
-  IL: stateOptions;
-  IN: stateOptions;
-  IA: stateOptions;
-  KS: stateOptions;
-  KY: stateOptions;
-  LA: stateOptions;
-  ME: stateOptions;
-  MD: stateOptions;
-  MA: stateOptions;
-  MI: stateOptions;
-  MN: stateOptions;
-  MS: stateOptions;
-  MO: stateOptions;
-  MT: stateOptions;
-  NE: stateOptions;
-  NV: stateOptions;
-  NH: stateOptions;
-  NJ: stateOptions;
-  NM: stateOptions;
-  NY: stateOptions;
-  NC: stateOptions;
-  ND: stateOptions;
-  OH: stateOptions;
-  OK: stateOptions;
-  OR: stateOptions;
-  PA: stateOptions;
-  PR: stateOptions;
-  RI: stateOptions;
-  SC: stateOptions;
-  SD: stateOptions;
-  TX: stateOptions;
-  UT: stateOptions;
-  VT: stateOptions;
-  VA: stateOptions;
-  WA: stateOptions;
-  WV: stateOptions;
-  WI: stateOptions;
-  WY: stateOptions;
-};
